auth_service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["Астраханцев Павел"]
package-mode = false 

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.30"}
python-jose = {extras = ["cryptography"], version = "^3.3"}
redis = "^5.0"
httpx = "^0.27"
pydantic = "^2.7"
pydantic-settings = "^2.3"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"